refactor: move utils to proper file
