+++ packages/commons/src/entity-event/tasklist-event.schemas.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod'
 import { ValuesOf } from '../utils'
-import { pickFromObj } from '../utils/zod-schema.utils'
+import { pickFromObj } from '../utils/object.utils'
 import { EntityEventType } from './entity-event.schemas'
 
 export const TasklistEventType = pickFromObj(EntityEventType, ['TitleChanged', 'ListParentListChanged'])
